requirements.txt
alembic==1.7.5
autopep8==1.6.0
click==7.1.2
Flask==2.0.2
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-Mobility==1.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
importlib-metadata==4.8.2
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
psycopg2==2.9.2
pycodestyle==2.8.0
python-dotenv==0.19.2
SQLAlchemy==1.4.27
toml==0.10.2
Werkzeug==2.0.2
WTForms==3.0.0
zipp==3.6.0
